move: 유저 dto 디렉토리 정리 (#CHZZ-80)
