+++ src/main/java/org/chzz/market/domain/auction/repository/AuctionRepositoryImpl.java
@@ -33,8 +33,8 @@ import org.chzz.market.domain.auction.dto.response.QAuctionResponse;
 import org.chzz.market.domain.auction.dto.response.QMyAuctionResponse;
 import org.chzz.market.domain.image.entity.QImage;
 import org.chzz.market.domain.product.entity.Product.Category;
-import org.chzz.market.domain.user.dto.ParticipationCountsResponse;
-import org.chzz.market.domain.user.dto.QParticipationCountsResponse;
+import org.chzz.market.domain.user.dto.response.ParticipationCountsResponse;
+import org.chzz.market.domain.user.dto.response.QParticipationCountsResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
