+++ src/main/java/org/chzz/market/domain/user/dto/response/ParticipationCountsResponse.java
@@ -1,4 +1,4 @@
-package org.chzz.market.domain.user.dto;
+package org.chzz.market.domain.user.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
