+++ src/main/java/org/chzz/market/domain/user/service/CustomOAuth2UserService.java
@@ -3,9 +3,9 @@ package org.chzz.market.domain.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.chzz.market.domain.user.dto.CustomUserDetails;
-import org.chzz.market.domain.user.dto.KaKaoResponse;
-import org.chzz.market.domain.user.dto.NaverResponse;
-import org.chzz.market.domain.user.dto.OAuth2Response;
+import org.chzz.market.domain.user.dto.response.KaKaoResponse;
+import org.chzz.market.domain.user.dto.response.NaverResponse;
+import org.chzz.market.domain.user.dto.response.OAuth2Response;
 import org.chzz.market.domain.user.entity.User;
 import org.chzz.market.domain.user.entity.User.ProviderType;
 import org.chzz.market.domain.user.repository.UserRepository;
