+++ src/main/java/org/chzz/market/domain/user/service/UserService.java
@@ -7,8 +7,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.chzz.market.domain.user.dto.request.UserCreateRequest;
 import org.chzz.market.domain.user.dto.response.NicknameAvailabilityResponse;
 import org.chzz.market.domain.auction.repository.AuctionRepository;
-import org.chzz.market.domain.user.dto.ParticipationCountsResponse;
-import org.chzz.market.domain.user.dto.UserProfileResponse;
+import org.chzz.market.domain.user.dto.response.ParticipationCountsResponse;
+import org.chzz.market.domain.user.dto.response.UserProfileResponse;
 import org.chzz.market.domain.user.entity.User;
 import org.chzz.market.domain.user.error.UserErrorCode;
 import org.chzz.market.domain.user.error.exception.UserException;
